fix warnings

+++ caldgemm_opencl.cpp
@@ -686,7 +686,7 @@ int caldgemm_opencl::ExecuteKernels(caldgemm::DGEMMPrepareAndExecuteTask& Task, 
 	size_t global_size[2] = {GROUP_SIZE_X * GROUP_COUNT_X, GROUP_SIZE_Y * GROUP_COUNT_Y};
 #else
 	size_t local_size[2] = {8, 8};
-	size_t global_size[2] = {height1 / 4, height2 / 4};
+	size_t global_size[2] = {(size_t) height1 / 4, (size_t) height2 / 4};
 #endif
 	if (Config->VerboseTiming)
 	{
@@ -712,7 +712,7 @@ int caldgemm_opencl::ExecuteKernels(caldgemm::DGEMMPrepareAndExecuteTask& Task, 
 		if (Config->GPU_C)
 		{
 			size_t origin[3] = {0, 0, 0};
-			size_t region[3] = {height1 * sizeof(double), height2, 1};
+			size_t region[3] = {(size_t) height1 * sizeof(double), (size_t) height2, 1};
 			if (Config->Debug) fprintf(STD_OUT, "Transfer C from GPU: region %d x %d\n", (int) region[0], (int) region[1]);
 			if (Config->ThreadSaveDriver == -1) pthread_mutex_lock(&globalDriverLock);
 			CHKRET(clEnqueueReadBufferRect(ocl_command_queues[Task.device][Task.j], ocl_cbuffers[Task.device][Task.j], CL_FALSE, origin, origin, region, 0, 0, C_pitch * sizeof(double), 0, C + blockn * Config->Height + blockm * Config->Height * C_pitch, 0, NULL, &ocl_events[Task.device][Task.j]), "Error retrieving C\n");
